LPS-58194 SF - move site administration application service to product navigation
