+++ modules/apps/product-navigation/product-navigation-site-administration/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <projectDescription>
-	<name>product-menu-site-administration-master</name>
+	<name>product-navigation-site-administration-master</name>
 	<comment></comment>
 	<projects></projects>
 	<buildSpec>
