+++ modules/apps/product-navigation/product-navigation-site-administration/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Product Menu Site Administration Service
+Bundle-Name: Liferay Product Navigation Site Administration Service
 Bundle-SymbolicName: com.liferay.product.menu.site.administration.service
 Bundle-Version: 1.0.0
 Include-Resource:\
