+++ modules/apps/product-navigation/product-navigation-site-administration/build.gradle
