+++ modules/apps/product-navigation/product-navigation-site-administration/snapshot.properties
