+++ modules/apps/product-navigation/product-navigation-site-administration/src/META-INF/resources/content/init.jsp
