+++ modules/apps/product-navigation/product-navigation-site-administration/src/META-INF/resources/my_sites/init.jsp
