+++ modules/apps/product-navigation/product-navigation-site-administration/src/com/liferay/product/navigation/site/administration/application/list/ContentPanelCategory.java
@@ -12,19 +12,13 @@
  * details.
  */
 
-package com.liferay.product.menu.site.administration.application.list;
+package com.liferay.product.navigation.site.administration.application.list;
 
 import com.liferay.application.list.BaseJSPPanelCategory;
 import com.liferay.application.list.PanelCategory;
 import com.liferay.application.list.constants.PanelCategoryKeys;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.model.Group;
-import com.liferay.portal.model.Organization;
-import com.liferay.portal.model.User;
-import com.liferay.portal.security.permission.PermissionChecker;
 
-import java.util.List;
 import java.util.Locale;
 
 import javax.servlet.ServletContext;
@@ -38,49 +32,31 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"panel.category.key=" + PanelCategoryKeys.SITES_ALL_SITES,
+		"panel.category.key=" + PanelCategoryKeys.SITE_ADMINISTRATION,
 		"service.ranking:Integer=200"
 	},
 	service = PanelCategory.class
 )
-public class MySitesPanelCategory extends BaseJSPPanelCategory {
+public class ContentPanelCategory extends BaseJSPPanelCategory {
 
 	@Override
 	public String getIconCssClass() {
-		return "icon-sitemap";
+		return "icon-file-text";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/META-INF/resources/my_sites/my_sites.jsp";
+		return "/META-INF/resources/content/content.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return PanelCategoryKeys.SITES_ADMINISTRATION_MY_SITES;
+		return PanelCategoryKeys.SITE_ADMINISTRATION_CONTENT;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "my-sites");
-	}
-
-	@Override
-	public boolean hasAccessPermission(
-			PermissionChecker permissionChecker, Group group)
-		throws PortalException {
-
-		User user = permissionChecker.getUser();
-
-		List<Group> siteGroups = user.getMySiteGroups(
-			new String[] {Group.class.getName(), Organization.class.getName()},
-			false, 1);
-
-		if (siteGroups.isEmpty()) {
-			return false;
-		}
-
-		return super.hasAccessPermission(permissionChecker, group);
+		return LanguageUtil.get(locale, "category.site_administration.content");
 	}
 
 	@Override
