+++ modules/apps/product-navigation/product-navigation-site-administration/src/com/liferay/product/navigation/site/administration/application/list/PagesPanelCategory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.product.menu.site.administration.application.list;
+package com.liferay.product.navigation.site.administration.application.list;
 
 import com.liferay.application.list.BasePanelCategory;
 import com.liferay.application.list.PanelCategory;
@@ -30,26 +30,25 @@ import org.osgi.service.component.annotations.Component;
 	immediate = true,
 	property = {
 		"panel.category.key=" + PanelCategoryKeys.SITE_ADMINISTRATION,
-		"service.ranking:Integer=400"
+		"service.ranking:Integer=100"
 	},
 	service = PanelCategory.class
 )
-public class ConfigurationPanelCategory extends BasePanelCategory {
+public class PagesPanelCategory extends BasePanelCategory {
 
 	@Override
 	public String getIconCssClass() {
-		return "icon-hdd";
+		return "icon-sitemap";
 	}
 
 	@Override
 	public String getKey() {
-		return PanelCategoryKeys.SITE_ADMINISTRATION_CONFIGURATION;
+		return PanelCategoryKeys.SITE_ADMINISTRATION_PAGES;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(
-			locale, "category.site_administration.configuration");
+		return LanguageUtil.get(locale, "category.site_administration.pages");
 	}
 
 }
