+++ modules/apps/product-navigation/product-navigation-site-administration/src/com/liferay/product/navigation/site/administration/application/list/PublishingToolsPanelCategory.java
@@ -12,17 +12,12 @@
  * details.
  */
 
-package com.liferay.product.menu.site.administration.application.list;
+package com.liferay.product.navigation.site.administration.application.list;
 
 import com.liferay.application.list.BasePanelCategory;
 import com.liferay.application.list.PanelCategory;
 import com.liferay.application.list.constants.PanelCategoryKeys;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.model.Group;
-import com.liferay.portal.security.permission.ActionKeys;
-import com.liferay.portal.security.permission.PermissionChecker;
-import com.liferay.portal.service.permission.PortalPermissionUtil;
 
 import java.util.Locale;
 
@@ -34,40 +29,27 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true,
 	property = {
-		"panel.category.key=" + PanelCategoryKeys.SITES_ALL_SITES,
-		"service.ranking:Integer=100"
+		"panel.category.key=" + PanelCategoryKeys.SITE_ADMINISTRATION,
+		"service.ranking:Integer=500"
 	},
 	service = PanelCategory.class
 )
-public class AdministrationPanelCategory extends BasePanelCategory {
+public class PublishingToolsPanelCategory extends BasePanelCategory {
 
 	@Override
 	public String getIconCssClass() {
-		return "icon-sitemap";
+		return "icon-retweet";
 	}
 
 	@Override
 	public String getKey() {
-		return PanelCategoryKeys.SITES_ADMINISTRATION_SITES;
+		return PanelCategoryKeys.SITE_ADMINISTRATION_PUBLISHING_TOOLS;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "administration");
-	}
-
-	@Override
-	public boolean hasAccessPermission(
-			PermissionChecker permissionChecker, Group group)
-		throws PortalException {
-
-		if (PortalPermissionUtil.contains(
-				permissionChecker, ActionKeys.VIEW_CONTROL_PANEL)) {
-
-			return true;
-		}
-
-		return false;
+		return LanguageUtil.get(
+			locale, "category.site_administration.publishing_tools");
 	}
 
 }
