+++ modules/apps/product-navigation/product-navigation-site-administration/src/com/liferay/product/navigation/site/administration/application/list/SiteAdministrationPanelCategory.java
@@ -12,27 +12,24 @@
  * details.
  */
 
-package com.liferay.product.menu.site.administration.application.list;
+package com.liferay.product.navigation.site.administration.application.list;
 
 import com.liferay.application.list.BaseJSPPanelCategory;
 import com.liferay.application.list.PanelCategory;
-import com.liferay.application.list.PanelCategoryRegistry;
 import com.liferay.application.list.constants.PanelCategoryKeys;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.model.Group;
+import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.security.permission.PermissionChecker;
+import com.liferay.portal.service.permission.GroupPermissionUtil;
 
-import java.util.List;
 import java.util.Locale;
 
 import javax.servlet.ServletContext;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
 
 /**
  * @author Eudaldo Alonso
@@ -40,12 +37,12 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
 @Component(
 	immediate = true,
 	property = {
-		"panel.category.key=" + PanelCategoryKeys.ROOT,
+		"panel.category.key=" + PanelCategoryKeys.SITES,
 		"service.ranking:Integer=100"
 	},
 	service = PanelCategory.class
 )
-public class SitesPanelCategory extends BaseJSPPanelCategory {
+public class SiteAdministrationPanelCategory extends BaseJSPPanelCategory {
 
 	@Override
 	public String getIconCssClass() {
@@ -54,17 +51,17 @@ public class SitesPanelCategory extends BaseJSPPanelCategory {
 
 	@Override
 	public String getJspPath() {
-		return "/META-INF/resources/sites/sites.jsp";
+		return "/META-INF/resources/sites/site_administration.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return PanelCategoryKeys.SITES;
+		return PanelCategoryKeys.SITE_ADMINISTRATION;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "category.sites");
+		return LanguageUtil.get(locale, "site-administration");
 	}
 
 	@Override
@@ -72,19 +69,14 @@ public class SitesPanelCategory extends BaseJSPPanelCategory {
 			PermissionChecker permissionChecker, Group group)
 		throws PortalException {
 
-		if (_panelCategoryRegistry == null) {
-			return false;
+		if (GroupPermissionUtil.contains(
+				permissionChecker, group,
+				ActionKeys.VIEW_SITE_ADMINISTRATION)) {
+
+			return true;
 		}
 
-		List<PanelCategory> childPanelCategories =
-			_panelCategoryRegistry.getChildPanelCategories(
-				this, permissionChecker, group);
-
-		if (childPanelCategories.isEmpty()) {
-			return false;
-		}
-
-		return super.hasAccessPermission(permissionChecker, group);
+		return false;
 	}
 
 	@Override
@@ -95,24 +87,5 @@ public class SitesPanelCategory extends BaseJSPPanelCategory {
 	public void setServletContext(ServletContext servletContext) {
 		super.setServletContext(servletContext);
 	}
-
-	@Reference(
-		cardinality = ReferenceCardinality.OPTIONAL,
-		policy = ReferencePolicy.DYNAMIC,
-		policyOption = ReferencePolicyOption.GREEDY
-	)
-	protected void setPanelCategoryRegistry(
-		PanelCategoryRegistry panelCategoryRegistry) {
-
-		_panelCategoryRegistry = panelCategoryRegistry;
-	}
-
-	protected void unsetPanelCategoryRegistry(
-		PanelCategoryRegistry panelCategoryRegistry) {
-
-		_panelCategoryRegistry = null;
-	}
-
-	private PanelCategoryRegistry _panelCategoryRegistry;
 
 }
