+++ modules/apps/product-navigation/product-navigation-site-administration/src/com/liferay/product/navigation/site/administration/application/list/SitesPanelCategory.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.product.menu.site.administration.application.list;
+package com.liferay.product.navigation.site.administration.application.list;
 
-import com.liferay.application.list.BasePanelCategory;
+import com.liferay.application.list.BaseJSPPanelCategory;
 import com.liferay.application.list.PanelCategory;
 import com.liferay.application.list.PanelCategoryRegistry;
 import com.liferay.application.list.constants.PanelCategoryKeys;
@@ -25,6 +25,8 @@ import com.liferay.portal.security.permission.PermissionChecker;
 
 import java.util.List;
 import java.util.Locale;
+
+import javax.servlet.ServletContext;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -38,12 +40,12 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
 @Component(
 	immediate = true,
 	property = {
-		"panel.category.key=" + PanelCategoryKeys.SITES,
-		"service.ranking:Integer=200"
+		"panel.category.key=" + PanelCategoryKeys.ROOT,
+		"service.ranking:Integer=100"
 	},
 	service = PanelCategory.class
 )
-public class AllSitesPanelCategory extends BasePanelCategory {
+public class SitesPanelCategory extends BaseJSPPanelCategory {
 
 	@Override
 	public String getIconCssClass() {
@@ -51,13 +53,18 @@ public class AllSitesPanelCategory extends BasePanelCategory {
 	}
 
 	@Override
+	public String getJspPath() {
+		return "/META-INF/resources/sites/sites.jsp";
+	}
+
+	@Override
 	public String getKey() {
-		return PanelCategoryKeys.SITES_ALL_SITES;
+		return PanelCategoryKeys.SITES;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "all-sites");
+		return LanguageUtil.get(locale, "category.sites");
 	}
 
 	@Override
@@ -78,6 +85,15 @@ public class AllSitesPanelCategory extends BasePanelCategory {
 		}
 
 		return super.hasAccessPermission(permissionChecker, group);
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.product.menu.site.administration.service)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
 	}
 
 	@Reference(
