+++ modules/apps/product-navigation/product-navigation-site-administration/src/com/liferay/product/navigation/site/administration/application/list/UsersPanelCategory.java
@@ -12,19 +12,16 @@
  * details.
  */
 
-package com.liferay.product.menu.site.administration.application.list;
+package com.liferay.product.navigation.site.administration.application.list;
 
-import com.liferay.application.list.BaseJSPPanelCategory;
+import com.liferay.application.list.BasePanelCategory;
 import com.liferay.application.list.PanelCategory;
 import com.liferay.application.list.constants.PanelCategoryKeys;
 import com.liferay.portal.kernel.language.LanguageUtil;
 
 import java.util.Locale;
 
-import javax.servlet.ServletContext;
-
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Eudaldo Alonso
@@ -33,39 +30,25 @@ import org.osgi.service.component.annotations.Reference;
 	immediate = true,
 	property = {
 		"panel.category.key=" + PanelCategoryKeys.SITE_ADMINISTRATION,
-		"service.ranking:Integer=200"
+		"service.ranking:Integer=300"
 	},
 	service = PanelCategory.class
 )
-public class ContentPanelCategory extends BaseJSPPanelCategory {
+public class UsersPanelCategory extends BasePanelCategory {
 
 	@Override
 	public String getIconCssClass() {
-		return "icon-file-text";
-	}
-
-	@Override
-	public String getJspPath() {
-		return "/META-INF/resources/content/content.jsp";
+		return "icon-group";
 	}
 
 	@Override
 	public String getKey() {
-		return PanelCategoryKeys.SITE_ADMINISTRATION_CONTENT;
+		return PanelCategoryKeys.SITE_ADMINISTRATION_USERS;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "category.site_administration.content");
-	}
-
-	@Override
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.product.menu.site.administration.service)",
-		unbind = "-"
-	)
-	public void setServletContext(ServletContext servletContext) {
-		super.setServletContext(servletContext);
+		return LanguageUtil.get(locale, "category.site_administration.users");
 	}
 
 }
